* master: (169 commits)
  Fix relative constant addressing in file cache
  Remove outdated TODO files
  Update coding standards
  ext/imap/config.m4: fix ac_cv_u8t_decompose check
  Fixed a typo
  Make ini formatting more consistent
  Make sizeof formatting consistent
  Remove RAND_RANGE() macro
  Implement list() reference assignments
  Move temporary allocation to the stack
  Fix tests
  Use cheaper ascii conversion for mode
  Revamp fopen implementation, rely on open
  Put missing signature into header
  Cleanup
  Rework mkdir impl and expose wide char variant
  Remove dead code
  Move declarations
  Improve ioutil access impl and refactor tsrm_win32_access
  Revamp unlink() implementation and improve error handling
  ...

+++ ext/standard/mt_rand.c
@@ -294,7 +294,7 @@ PHPAPI zend_long php_mt_rand_range(zend_long min, zend_long max)
  * rand() allows min > max, mt_rand does not */
 PHPAPI zend_long php_mt_rand_common(zend_long min, zend_long max)
 {
-	uint32_t n;
+	int64_t n;
 
 	if (BG(mt_rand_mode) == MT_RAND_MT19937) {
 		return php_mt_rand_range(min, max);
@@ -302,7 +302,7 @@ PHPAPI zend_long php_mt_rand_common(zend_long min, zend_long max)
 
 	/* Legacy mode deliberately not inside php_mt_rand_range()
 	 * to prevent other functions being affected */
-	n = php_mt_rand() >> 1;
+	n = (int64_t)php_mt_rand() >> 1;
 	RAND_RANGE_BADSCALING(n, min, max, PHP_MT_RAND_MAX);
 
 	return n;
